+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.6 - 2024-02-26
+
+### Changed
+
+- Split RequestRule to two separate classes for block and unblock rules
+- Refactor some strings to consts
+
+
 ## 1.2.5 - 2024-02-16
 
 ### Changed